Resolve race conditions with R/W lock in Server
Running the tests in this repo with the `-race` flag detects multiple
race conditions.

First course of action is to add a `sync.Mutex` to the `Server` type,
then replace direct reads and writes to `isStarted` and `PortNumber`
fields with setter and getter methods which lock until it is safe to
access these values. Unfortunately, to ensure safety, this necessitates
a breaking change - the `PortNumber` field cannot be exposed any longer,
so it is made a private field and users must call the `PortNumber()`
method instead.

Secondly, it is not safe for the `Messages()` method to return a slice
of pointers to Message data, this has been changed to return a copy of
the data instead.