Improve email regex
In the RFC, these characters are allowed in the local part. I was
using this for testing and found that `+` didn't work, however that's a
pretty commonly used character in email addresses. When I looked into
the spec I found that there were many characters that are valid and so I
added them to the regex.